+++ src/components/Profiling/Dialog/ProfilingWizard.tsx
@@ -8,8 +8,13 @@ import React, { useEffect, useRef } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { appendFile, existsSync, mkdirSync } from 'fs';
 import path from 'path';
+import { ConfirmationDialog } from 'pc-nrfconnect-shared';
 
 import { RootState } from '../../../appReducer';
+import {
+    addConfirmBeforeClose,
+    clearConfirmBeforeClose,
+} from '../../../features/confirmBeforeClose/confirmBeforeCloseSlice';
 import { startProcessingCsv } from '../../../features/nrfutillNpm/csvProcessing';
 import {
     getBucks,
@@ -21,6 +26,9 @@ import {
     setEventRecordingPath,
 } from '../../../features/pmicControl/pmicControlSlice';
 import {
+    clearAbortAction,
+    closeProfiling,
+    getAbort,
     getCcProfilingState,
     getProfile,
     getProfileIndex,
@@ -76,8 +84,26 @@ export default () => {
     const fuelGauge = useSelector(getFuelGauge);
     const index = useSelector(getProfileIndex);
     const ccProfilingState = useSelector(getCcProfilingState);
+    const abortAction = useSelector(getAbort);
 
     const dispatch = useDispatch();
+
+    useEffect(() => {
+        if (profilingStage) {
+            dispatch(
+                addConfirmBeforeClose({
+                    id: 'PROFILING_WIZARD',
+                    message: 'Profiling is currently ongoing.',
+                    onClose() {
+                        dispatch(closeProfiling());
+                    },
+                })
+            );
+            return () => {
+                dispatch(clearConfirmBeforeClose('PROFILING_WIZARD'));
+            };
+        }
+    }, [dispatch, profilingStage]);
 
     useEffect(() => {
         if (profilingStage === 'Charging') {
@@ -270,16 +296,37 @@ export default () => {
     }, [ccProfilingState, dispatch]);
 
     return (
-        <>
-            {profilingStage === 'MissingSyncBoard' && (
-                <PreConfigurationDialog />
+        <div>
+            {abortAction ? (
+                <ConfirmationDialog
+                    title="Aborting profiling"
+                    isVisible
+                    onConfirm={() => {
+                        abortAction();
+                        dispatch(clearAbortAction());
+                    }}
+                    onCancel={() => {
+                        dispatch(clearAbortAction());
+                    }}
+                >
+                    Any unfinished processes will be lost. Are you sure you want
+                    to abort profiling?
+                </ConfirmationDialog>
+            ) : (
+                <>
+                    {profilingStage === 'MissingSyncBoard' && (
+                        <PreConfigurationDialog />
+                    )}
+                    {profilingStage === 'Configuration' && (
+                        <ConfigurationDialog />
+                    )}
+                    {profilingStage === 'Checklist' && <ChecklistDialog />}
+                    {profilingStage === 'Charging' && <ChargingDialog />}
+                    {profilingStage === 'Resting' && <RestingDialog />}
+                    {profilingStage === 'Profiling' && <ProfilingDialog />}
+                    {profilingStage === 'Complete' && <CompleteDialog />}
+                </>
             )}
-            {profilingStage === 'Configuration' && <ConfigurationDialog />}
-            {profilingStage === 'Checklist' && <ChecklistDialog />}
-            {profilingStage === 'Charging' && <ChargingDialog />}
-            {profilingStage === 'Resting' && <RestingDialog />}
-            {profilingStage === 'Profiling' && <ProfilingDialog />}
-            {profilingStage === 'Complete' && <CompleteDialog />}
-        </>
+        </div>
     );
 };