+++ Changelog.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.7.2 - 2023-06-05
+
+### Added
+
+-   Confirm dialog aborting profiling
+-   Confirm dialog before app closes while profiling
+-   Confirm dialog before app closes while data is processing
+
 ## 0.7.1 - 2023-06-02
 
 ### Added